requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.31
pydantic==2.8.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.3.0
httpx==0.28.1
streamlit==1.37.1
plotly==5.17.0

# Documents & processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.2
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.10

# AI/ML
transformers>=4.45.0,<5
tokenizers>=0.20.1
torch==2.9.0
sentence-transformers==3.0.1

# Vector search
chromadb==0.4.18

# Infra & misc
alembic==1.13.1
redis==5.0.1
celery==5.3.4
jinja2==3.1.2

# Testing
pytest==8.3.2

# Dev tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1